backend/requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
Flask==3.1.2
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
gevent==25.9.1
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.1.0
marshmallow-sqlalchemy==1.4.2
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==6.1
zope.interface==8.1.1
